Fixing member type issue

+++ src/vehicle.h
@@ -73,7 +73,7 @@ public:
 
   void configure(vector<float> road_data);
 
-  float position_at(int t);
+  float position_at(double t);
 
   bool get_vehicle_behind(map<int, vector<Vehicle>> predictions, int lane, Vehicle & rVehicle);
